Project structure refractor and rice enc/dec
