+++ include/test.h
@@ -1,0 +1,19 @@
+#ifndef TEST_H
+#define TEST_H
+
+#include "../include/precomp.h"
+
+void print_hash(unsigned char *);
+void hash_n(uint32_t);
+void test_reduction();
+void test_hash_reduction();
+void test_initialize();
+void test_generate();
+void test_swap();
+void test_sort();
+void test_gen_sort();
+void test_clean();
+void test_rice();
+void test_ceri();
+
+#endif
